moved repository-search function into GitHubAPI
